requirements.txt
langchain
langgraph
langchain_community
langchain_core
langchain_groq
langchain_openai
faiss-cpu
streamlit
tavily-python
langchain-tavily
python-dotenv
logtail-python
langchain_experimental
sentence-transformers
numpy
scikit-learn
scipy
pandas
matplotlib
seaborn
plotly
nltk
spacy

